4/20

+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,9 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
 
 public interface IDamageable
 {
-    void TakeDamage(float damage, bool isCritical = false);
+    void TakeDamage(int damage, bool isCritical = false);
     void Die();
     bool IsAlive { get; }
-    float CurrentHealth { get; }
-    float MaxHealth { get; }
+    int CurrentHealth { get; }
+    int MaxHealth { get; }
 }
